Fix missing producer config (CORE-550)
The CQRS module created KafkaProducer's only with the bootstrap.servers
from the Kafka Connector meaning that if additional configuration was
necessary, e.g. for a secure cluster, this was ignored and the producers
would get stuck in a fail loop spamming the log with WARN messages
about failures to connect to Kafka.  This is fixed to now simply pass
the configuration across as a producer will ignore the consumer specific
configuration it doesn't use but does need to share generic
configuratio, like AuthN.