Avoid unnecessary compactions
This commit resolves an intermittently seen issue in dev clusters where
sometimes after restart when performing the initial compaction we can
get into a crash/restart loop as compaction impacts responsiveness of
the service.  Also, if the dataset is already compacted this compaction
is unnecessary.

To avoid this we now store a .last-compaction file in the database
directory which allows us to compare the current size with the
previously compacted size to determine if the database has grown since
last compaction.  If it hasn't grown we can skip the initial compaction
entirely.

Related to this spotted and fixed a bug where the determination of
database size was incorrectly sizing the entire database directory,
rather than just the active Data-NNNN directory.  We now only size the
current Data-NNNN directory which avoids sizing inactive database files
and other metadata files in the root database directory.