Fix missing roles policy for /$/compact/<dataset> (CORE-1042)
While working on integration testing it was noted that attempting to
call the /$/compact/<dataset> endpoint was failing with a Denied to All
Users error.  Upon investigation it was realised that a suitable roles
policy was not in place for this endpoint so it always hit the fallack
Deny All policy.  This commit fixes that omission.

Also some minor improvements to the docker-run.sh script to disable
build caching to ensure locally built images always have fresh code, and
to allow a DOCKER_RUN_OPTS environment variable to make it easier to
inject custom environment variables without modifying the script

Co-authored-by: Karolina Kadzielawa <karolina.kadzielawa@telicent.io>