Upgrade to Fuseki Kafka 2.2.0
This picks up a bug fix to use batch size in bytes as well as number of
events as primary batch commit triggers.  This avoids a potential OOM
that can occur if events are large because it tried to buffer too many
events in memory before committing.  Batches are now committed when
either the batch size in bytes, or the number of events, is reached.

This also includes improvements to Kafka related logging